% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edge_list_110}
\alias{edge_list_110}
\title{Edge list of cosponsorships of the House of Representatives of the 110th Congress}
\format{A dataframe with 177028 rows and 2 variables
\describe{
  \item{bill}{bill(arbitrary designation number)}
  \item{member}{Member of Congress}
}}
\source{
\url{http://jhfowler.ucsd.edu/cosponsorship.htm}
}
\usage{
edge_list_110
}
\description{
Edge list of cosponsorships of the House of Representatives of the 110th Congress
}
\keyword{datasets}
